requirements.txt
autopep8>=2.3.1
beautifulsoup4>=4.12.3
certifi>=2024.8.30
cffi>=1.17.1
charset-normalizer>=3.4.0
colorgram.py>=1.2.0
cssselect2>=0.7.0
defusedxml>=0.7.1
idna>=3.10
imageio>=2.36.0
numpy>=2.1.0
Pillow>=11.0.0
pycodestyle>=2.12.1
pycparser>=2.22
qrcode[pil]>=8.0
requests>=2.32.3
soupsieve>=2.6
tinycss2>=1.4.0
toml>=0.10.2
urllib3>=2.2.3
webencodings>=0.5.1
